Update password escaping for kumo POST request.
It appears that the HomeAssistant flow simply passes the user password to the pykumo library as a string after validation by voluptuous. When the password is interpolated into the body and included in the POST, it can fail due to un-escaped characters. In order to address this, the body is constructed as a regular dictionary and the json.dumps method is used to set it into the request. This has the benefit of escaping any necessary characters in the password field.